ingestor/requirements.txt
requests>=2.31.0
requests-oauthlib>=1.3.1
beautifulsoup4>=4.12.2
feedparser>=6.0.10
supabase>=2.15.1
httpx>=0.26.0
python-dotenv>=1.0.0
openai>=1.0.0
websockets==14.2
